crawler/requirement.txt
PyYAML==3.11
progress==1.2